Update Clock_with_alarm.c
This's a simple code, in C. I did it to practice. It has a simple function, you'll put the atual hour and minute, and it'll count each seconds of the time passing, print it and clear the screen after 1second, when it count 60 second=1min the variable second get 0 in value and mins get +1, and it repeat with mins and hour, this's very simple.